refactor(billing): move the post-commit collector to internal/postcommit and warn when the inline fallback is taken
